+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Bump kotlin to v1.6.21
 - Bump Lottie to v5.1.1
+- Migrate `DeletePatientEffectHandler` to use view effects
 
 ### Features